linux-gen: init: adding configuration file parsing
The parsing of the odp.conf configuration file is added.
The file is searched first in filename specified in the environment
variable $ODP_SYSCONFIG_FILE (where the special string "none" prevent
any file loading)
The file is then searched in the user home directory (~) and
then is the $prefix/etc directory.
This requires libconfig (sudo apt-get install libconfig-dev)

Signed-off-by: Christophe Milard <christophe.milard@linaro.org>
Reviewed-and-tested-by: Yi He <yi.he@linaro.org>
Signed-off-by: Maxim Uvarov <maxim.uvarov@linaro.org>